+++ eras/dijkstra/impl/src/Cardano/Ledger/Dijkstra/Rules/Epoch.hs
@@ -51,8 +51,9 @@ import Cardano.Ledger.Conway.Rules (
   updateNumDormantEpochs,
  )
 import Cardano.Ledger.Conway.State
-import Cardano.Ledger.Dijkstra.Era (EPOCH, SNAP)
+import Cardano.Ledger.Dijkstra.Era (EPOCH, POOLREAP, SNAP)
 import Cardano.Ledger.Dijkstra.PParams (DijkstraEraPParams)
+import Cardano.Ledger.Dijkstra.Rules.PoolReap ()
 import Cardano.Ledger.Dijkstra.Rules.Snap ()
 import Cardano.Ledger.Shelley.LedgerState (
   EpochState (..),
@@ -244,10 +245,10 @@ epochTransition = do
 
 instance
   ( Era era
-  , STS (Shelley.POOLREAP era)
+  , STS (POOLREAP era)
   , Event (EraRule "POOLREAP" era) ~ Shelley.ShelleyPoolreapEvent era
   ) =>
-  Embed (Shelley.POOLREAP era) (EPOCH era)
+  Embed (POOLREAP era) (EPOCH era)
   where
   wrapFailed = \case {}
   wrapEvent = PoolReapEvent